Move dto to new package
