+++ src/main/java/com/k9club/api/dto/user/OwnerUpdateDto.java
@@ -1,4 +1,4 @@
-package com.k9club.api.dto;
+package com.k9club.api.dto.user;
 
 public class OwnerUpdateDto extends UserBasicUpdateDto {
